pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atomic-red-team-mcp"
version = "1.2.3"
description = "MCP server providing access to Atomic Red Team security tests with search, validation, and creation capabilities"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "cyberbuff" }
]
keywords = [
    "mcp",
    "atomic-red-team"
]
dependencies = [
    "gitpython>=3.1.0",
    "pyyaml>=6.0.2",
    "pydantic>=2.0.0",
    "fastmcp>=2.12.4",
    "atomic-operator>=0.8.5",
]

[project.optional-dependencies]
dev = [
    "bump-my-version>=0.26.1",
]

[project.urls]
Homepage = "https://github.com/cyberbuff/atomic-red-team-mcp"
Documentation = "https://github.com/cyberbuff/atomic-red-team-mcp"
Repository = "https://github.com/cyberbuff/atomic-red-team-mcp"
Issues = "https://github.com/cyberbuff/atomic-red-team-mcp/issues"

[project.scripts]
atomic-red-team-mcp = "atomic_red_team_mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/atomic_red_team_mcp"]
exclude = [
    "/__pycache__",
    "/atomics",
    "/.git",
    "/.github",
    "/docker-compose.yaml",
    "/Dockerfile",
    "/.dockerignore",
    "/tests",
    "/docs",
    "/scripts",
    "/Makefile",
]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
    "/pyproject.toml",
]

[tool.bumpversion]
current_version = "1.2.3"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
tag = true
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
commit_args = ""

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "src/atomic_red_team_mcp/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'
